# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tektii Gateway API",
    "description": "Unified trading gateway API for multiple providers",
    "license": {
      "name": "Elastic License 2.0",
      "url": "https://www.elastic.co/licensing/elastic-license"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/v1/orders": {
      "get": {
        "tags": [
          "orders"
        ],
        "summary": "List open orders.",
        "operationId": "get_orders",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "description": "Filter by symbol",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "status",
            "in": "query",
            "description": "Filter by status (multiple allowed)",
            "required": false,
            "schema": {
              "type": [
                "array",
                "null"
              ],
              "items": {
                "$ref": "#/components/schemas/OrderStatus"
              }
            }
          },
          {
            "name": "side",
            "in": "query",
            "description": "Filter by side",
            "required": false,
            "schema": {
              "oneOf": [
                {
                  "type": "null"
                },
                {
                  "$ref": "#/components/schemas/Side"
                }
              ]
            }
          },
          {
            "name": "client_order_id",
            "in": "query",
            "description": "Filter by client order ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "oco_group_id",
            "in": "query",
            "description": "Filter by OCO group ID",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "name": "since",
            "in": "query",
            "description": "Filter orders created after this time",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "until",
            "in": "query",
            "description": "Filter orders created before this time",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ],
              "format": "date-time"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of orders to return",
            "required": false,
            "schema": {
              "type": [
                "integer",
                "null"
              ],
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid query parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "orders"
        ],
        "summary": "Submit a new order.",
        "description": "Returns an order handle with the assigned order ID.",
        "operationId": "submit_order",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrderRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Order submitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderHandle"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "422": {
            "description": "Order rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "501": {
            "description": "Operation not supported by provider",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "orders"
        ],
        "summary": "Cancel all orders.",
        "description": "Cancels all open orders, optionally filtered by symbol.",
        "operationId": "cancel_all_orders",
        "parameters": [
          {
            "name": "symbol",
            "in": "query",
            "description": "Filter by symbol (None = cancel all orders)",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Orders cancelled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CancelAllResult"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/positions/{position_id}": {
      "get": {
        "tags": [
          "positions"
        ],
        "summary": "Get position by ID.",
        "operationId": "get_position",
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "description": "Position ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Position details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Position"
                }
              }
            }
          },
          "404": {
            "description": "Position not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "positions"
        ],
        "summary": "Close a position.",
        "operationId": "close_position",
        "parameters": [
          {
            "name": "position_id",
            "in": "path",
            "description": "Position ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Optional close parameters",
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "type": "null"
                  },
                  {
                    "$ref": "#/components/schemas/ClosePositionRequest"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Position close order created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrderHandle"
                }
              }
            }
          },
          "404": {
            "description": "Position not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "503": {
            "description": "Provider unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "description": "API error response body (JSON wire format).",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "$ref": "#/components/schemas/ErrorCode",
            "description": "Machine-readable error code (e.g., `ORDER_NOT_FOUND`, `RATE_LIMITED`)"
          },
          "details": {
            "description": "Additional error context (varies by error type)"
          },
          "message": {
            "type": "string",
            "description": "Human-readable error message"
          }
        }
      },
...
      "CancelAllResult": {
        "type": "object",
        "description": "Result of cancelling multiple orders.",
        "required": [
          "cancelled_count",
          "failed_count"
        ],
        "properties": {
          "cancelled_count": {
            "type": "integer",
            "format": "int32",
            "description": "Number of orders successfully cancelled",
            "minimum": 0
          },
          "failed_count": {
            "type": "integer",
            "format": "int32",
            "description": "Number of orders that failed to cancel",
            "minimum": 0
          },
          "failed_order_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "IDs of orders that failed to cancel"
          }
        }
      },
...
      "ClosePositionRequest": {
        "type": "object",
        "description": "Request to close a position.",
        "properties": {
          "cancel_associated_orders": {
            "type": "boolean",
            "description": "Cancel pending SL/TP orders for this position (default: true)"
          },
          "limit_price": {
            "type": [
              "string",
              "null"
            ],
            "description": "Limit price (required if `order_type` is LIMIT)"
          },
          "order_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/OrderType",
                "description": "Order type for close (default: MARKET)"
              }
            ]
          },
          "quantity": {
            "type": [
              "string",
              "null"
            ],
            "description": "Quantity to close (None = close entire position)"
          }
        }
      },
...
      "ErrorCode": {
        "type": "string",
        "description": "Machine-readable error codes for API error responses.",
        "enum": [
          "INVALID_REQUEST",
          "INVALID_VALUE",
          "VALIDATION_ERROR",
          "UNAUTHORIZED",
          "ORDER_NOT_FOUND",
          "POSITION_NOT_FOUND",
          "SYMBOL_NOT_FOUND",
          "OCO_GROUP_NOT_FOUND",
          "ORDER_NOT_MODIFIABLE",
          "RESET_COOLDOWN",
          "ORDER_REJECTED",
          "RATE_LIMITED",
          "INTERNAL_ERROR",
          "UNSUPPORTED_OPERATION",
          "PROVIDER_ERROR",
          "SHUTTING_DOWN",
          "PROVIDER_UNAVAILABLE"
        ]
      },
...
      "MarginMode": {
        "type": "string",
        "description": "Margin mode for leveraged trading",
        "enum": [
          "CROSS",
          "ISOLATED"
        ]
      },
...
      "Order": {
        "type": "object",
        "description": "Full order details.",
        "required": [
          "id",
          "symbol",
          "side",
          "order_type",
          "quantity",
          "filled_quantity",
          "remaining_quantity",
          "status",
          "time_in_force",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "average_fill_price": {
            "type": [
              "string",
              "null"
            ],
            "description": "Weighted average fill price"
          },
          "client_order_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Client-provided order ID (if any)"
          },
          "correlation_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Gateway-assigned correlation ID for tracing this order through its lifecycle."
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the order was created"
          },
          "display_quantity": {
            "type": [
              "string",
              "null"
            ],
            "description": "Display quantity for iceberg orders"
          },
          "filled_quantity": {
            "type": "string",
            "description": "Quantity filled so far"
          },
          "hidden": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether this is a hidden order"
          },
          "id": {
            "type": "string",
            "description": "Server-assigned order ID"
          },
          "limit_price": {
            "type": [
              "string",
              "null"
            ],
            "description": "Limit price (for LIMIT and `STOP_LIMIT` orders)"
          },
          "oco_group_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "OCO group ID if this order is part of an OCO pair."
          },
          "order_type": {
            "$ref": "#/components/schemas/OrderType",
            "description": "Order type"
          },
          "position_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Associated position ID (for hedging accounts)"
          },
          "post_only": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether this is a post-only order"
          },
          "quantity": {
            "type": "string",
            "description": "Original order quantity"
          },
          "reduce_only": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "Whether this is a reduce-only order"
          },
          "reject_reason": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RejectReason",
                "description": "Rejection reason (if status is REJECTED)"
              }
            ]
          },
          "remaining_quantity": {
            "type": "string",
            "description": "Quantity remaining"
          },
          "side": {
            "$ref": "#/components/schemas/Side",
            "description": "Order side"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus",
            "description": "Current order status"
          },
          "stop_loss": {
            "type": [
              "string",
              "null"
            ],
            "description": "Stop loss price (for bracket orders)"
          },
          "stop_price": {
            "type": [
              "string",
              "null"
            ],
            "description": "Stop price (for STOP and `STOP_LIMIT` orders)"
          },
          "symbol": {
            "type": "string",
            "description": "Trading symbol"
          },
          "take_profit": {
            "type": [
              "string",
              "null"
            ],
            "description": "Take profit price (for bracket orders)"
          },
          "time_in_force": {
            "$ref": "#/components/schemas/TimeInForce",
            "description": "Time in force"
          },
          "trailing_distance": {
            "type": [
              "string",
              "null"
            ],
            "description": "Trailing distance (for trailing stop orders)"
          },
          "trailing_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TrailingType",
                "description": "Trailing type (absolute or percent)"
              }
            ]
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the order was last updated"
          }
        }
      },
...
      "OrderHandle": {
        "type": "object",
        "description": "Minimal response after order submission.",
        "required": [
          "id",
          "status"
        ],
        "properties": {
          "client_order_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Client-provided order ID (if any)"
          },
          "correlation_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Gateway-assigned correlation ID for tracing this order through its lifecycle."
          },
          "id": {
            "type": "string",
            "description": "Server-assigned order ID"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus",
            "description": "Initial order status"
          }
        }
      },
...
      "OrderRequest": {
        "type": "object",
        "description": "Request to submit a new order.",
        "required": [
          "symbol",
          "side",
          "quantity",
          "order_type"
        ],
        "properties": {
          "client_order_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Client-provided order ID for idempotency"
          },
          "display_quantity": {
            "type": [
              "string",
              "null"
            ],
            "description": "Visible quantity for iceberg orders"
          },
          "hidden": {
            "type": "boolean",
            "description": "If true, order not visible in order book"
          },
          "leverage": {
            "type": [
              "string",
              "null"
            ],
            "description": "Leverage multiplier"
          },
          "limit_price": {
            "type": [
              "string",
              "null"
            ],
            "description": "Required for LIMIT and `STOP_LIMIT` orders"
          },
          "margin_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MarginMode",
                "description": "Margin mode (cross or isolated)"
              }
            ]
          },
          "oco_group_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "OCO group ID to link orders (one-cancels-other)."
          },
          "order_type": {
            "$ref": "#/components/schemas/OrderType",
            "description": "Order type"
          },
          "position_id": {
            "type": [
              "string",
              "null"
            ],
            "description": "Target position ID (for hedging accounts)"
          },
          "post_only": {
            "type": "boolean",
            "description": "If true, reject if would immediately fill (maker only)"
          },
          "quantity": {
            "type": "string",
            "description": "Quantity in base units (must be positive)"
          },
          "reduce_only": {
            "type": "boolean",
            "description": "If true, can only reduce position"
          },
          "side": {
            "$ref": "#/components/schemas/Side",
            "description": "Buy or sell"
          },
          "stop_loss": {
            "type": [
              "string",
              "null"
            ],
            "description": "Stop loss price for bracket order"
          },
          "stop_price": {
            "type": [
              "string",
              "null"
            ],
            "description": "Required for STOP and `STOP_LIMIT` orders"
          },
          "symbol": {
            "type": "string",
            "description": "Canonical symbol format (e.g., \"EUR/USD\", \"AAPL\")"
          },
          "take_profit": {
            "type": [
              "string",
              "null"
            ],
            "description": "Take profit price for bracket order"
          },
          "time_in_force": {
            "$ref": "#/components/schemas/TimeInForce",
            "description": "Time in force (default: GTC)"
          },
          "trailing_distance": {
            "type": [
              "string",
              "null"
            ],
            "description": "Distance for `TRAILING_STOP` orders"
          },
          "trailing_type": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/TrailingType",
                "description": "Type of trailing stop (absolute or percent)"
              }
            ]
          }
        }
      },
      "OrderStatus": {
        "type": "string",
        "description": "Order status",
        "enum": [
          "PENDING",
          "PENDING_CANCEL",
          "OPEN",
          "PARTIALLY_FILLED",
          "FILLED",
          "CANCELLED",
          "REJECTED",
          "EXPIRED"
        ]
      },
      "OrderType": {
        "type": "string",
        "description": "Order type",
        "enum": [
          "MARKET",
          "LIMIT",
          "STOP",
          "STOP_LIMIT",
          "TRAILING_STOP"
        ]
      },
...
      "Position": {
        "type": "object",
        "description": "Position details.",
        "required": [
          "id",
          "symbol",
          "side",
          "quantity",
          "average_entry_price",
          "current_price",
          "unrealized_pnl",
          "realized_pnl",
          "opened_at",
          "updated_at"
        ],
        "properties": {
          "average_entry_price": {
            "type": "string",
            "description": "Weighted average entry price"
          },
          "current_price": {
            "type": "string",
            "description": "Current market price"
          },
          "id": {
            "type": "string",
            "description": "Position ID (may be synthetic for stock brokers)"
          },
          "leverage": {
            "type": [
              "string",
              "null"
            ],
            "description": "Leverage multiplier"
          },
          "liquidation_price": {
            "type": [
              "string",
              "null"
            ],
            "description": "Liquidation price (for leveraged positions)"
          },
          "margin_mode": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/MarginMode",
                "description": "Margin mode (for leveraged positions)"
              }
            ]
          },
          "opened_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the position was opened"
          },
          "quantity": {
            "type": "string",
            "description": "Current position quantity"
          },
          "realized_pnl": {
            "type": "string",
            "description": "Realized profit/loss (from partial closes)"
          },
          "side": {
            "$ref": "#/components/schemas/PositionSide",
            "description": "Position direction (long or short)"
          },
          "symbol": {
            "type": "string",
            "description": "Trading symbol"
          },
          "unrealized_pnl": {
            "type": "string",
            "description": "Unrealized profit/loss"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "description": "When the position was last modified"
          }
        }
      },
...
      "PositionSide": {
        "type": "string",
        "description": "Position side (long or short)",
        "enum": [
          "LONG",
          "SHORT"
        ]
      },
...
      "RejectReason": {
        "type": "string",
        "description": "Order rejection reason",
        "enum": [
          "INSUFFICIENT_MARGIN",
          "INSUFFICIENT_BALANCE",
          "INVALID_QUANTITY",
          "INVALID_PRICE",
          "SYMBOL_NOT_TRADEABLE",
          "MARKET_CLOSED",
          "REDUCE_ONLY_VIOLATED",
          "POST_ONLY_VIOLATED",
          "RATE_LIMITED",
          "UNKNOWN"
        ]
      },
      "Side": {
        "type": "string",
        "description": "Order side (buy or sell)",
        "enum": [
          "BUY",
          "SELL"
        ]
      },
      "TimeInForce": {
        "type": "string",
        "description": "Time in force for orders.",
        "enum": [
          "GTC",
          "DAY",
          "IOC",
          "FOK"
        ]
      },
...
      "TrailingType": {
        "type": "string",
        "description": "Trailing stop type (how trailing distance is interpreted)",
        "enum": [
          "ABSOLUTE",
          "PERCENT"
        ]
      },
...
    }
  },
...
}